REFACTOR loadingSpinner component
